pyproject.toml
[project]
name = "lab-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "jinja2>=3.1.6",
    "pyjwt>=2.10.1",
]
